Tweak typing
