+++ homeassistant/components/device_automation/condition.py
@@ -8,7 +8,7 @@ import voluptuous as vol
 from homeassistant.const import CONF_DOMAIN
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import config_validation as cv
-from homeassistant.helpers.condition import trace_condition_function
+from homeassistant.helpers.condition import ConditionProtocol, trace_condition_function
 from homeassistant.helpers.typing import ConfigType
 
 from . import DeviceAutomationType, async_get_device_automation_platform
@@ -18,23 +18,12 @@ if TYPE_CHECKING:
     from homeassistant.helpers import condition
 
 
-class DeviceAutomationConditionProtocol(Protocol):
+class DeviceAutomationConditionProtocol(ConditionProtocol, Protocol):
     """Define the format of device_condition modules.
 
-    Each module must define either CONDITION_SCHEMA or async_validate_condition_config.
+    Each module must define either CONDITION_SCHEMA or async_validate_condition_config
+    from ConditionProtocol.
     """
-
-    CONDITION_SCHEMA: vol.Schema
-
-    async def async_validate_condition_config(
-        self, hass: HomeAssistant, config: ConfigType
-    ) -> ConfigType:
-        """Validate config."""
-
-    def async_condition_from_config(
-        self, hass: HomeAssistant, config: ConfigType
-    ) -> condition.ConditionCheckerType:
-        """Evaluate state based on configuration."""
 
     async def async_get_condition_capabilities(
         self, hass: HomeAssistant, config: ConfigType
